component buttons will smooth scroll to list view on click
